Testando

+++ src/controllers/urls.controllers.js
@@ -82,6 +82,13 @@ export async function deleteUrlId (req, res,) {
 
         const url = await db.query(`SELECT * FROM encurtar WHERE id=$1`,[id])
         if (url.rows.length === 0) res.sendStatus(404)
+
+        const linkCreatorId = url.rows[0].criadorDaUrl
+        const sessionUserId = sessao.rows[0].idUsuario; 
+
+        if (linkCreatorId !== sessionUserId) {
+         return res.sendStatus(401);
+        }
        
         await db.query(`DELETE FROM encurtar WHERE id=$1`, [id])
